requirements.txt
httpx==0.27.0
python-telegram-bot==21.3
python-dotenv==1.0.1